Move user and auth interfaces and classes to its own file
